Autonomous Update v0.1
I have begun the reconstruction of autonomous. The basic looping stuff is in auto_independent now (could probably be a separate VI, but whatever.) I also put in my first command, which "theoretically" will have the robot drive to a specified distance. To get it to work, I need the gear ratio from the drive transmissions, and then tune it/test it. But it's written now, and likely only needs minor changes. I decided I'm not going to attempt motion profiling until we put the drive Spark MAXes on CAN - and boy do we need to.