Teleoperated Auto-Targeting
I've integrated the limelight into the teleop software on this one. In theory, when disconnected, nothing should change. When connected, holding right bumper on the driver joystick will autonomously turn the robot to the upper hub target and spool up the shooter when the robot is within range. Sorry there haven't been many updates on here, a lot of dev work has gone into a separate project due to Cobra's drive motors being on PWM instead of CAN.